Add batch learning
Most of it was already there, only the circular convolution had to be
updated. Shift vector can still only be -1, 0 or 1 - this limitation
remained.

Tested: although the result seems a bit more clunky compared to the
reference, not sure if it's the result of the changes or just random.
The reason in this case is probably that the NTM is fed batch_size * iteration
number of elements (in my test, 3 mil samples instead of 100k), so the previous,
cleaner result might be simply due to "early stopping". The test result is still
perfect on the provided sample, despite the unnecessary write operations, so I
consider it working for now. Further testing should be done anyway.